feat(build): move code used by native framework to separate lib
Move util code to separate static lib to be used in cocaine-native-framework.
This is needed to drop dependency of cocaine framework on libcocaine-core.
