+++ debian/libcocaine-dev.install
@@ -1,1 +1,2 @@
 usr/include/cocaine/*
+usr/lib/libcocaine-util.a
