+++ src/service/locator.cpp
@@ -748,49 +748,3 @@ locator_t::on_context_shutdown() {
 
     m_signals = nullptr;
 }
-
-namespace {
-
-// Locator errors
-
-struct locator_category_t:
-    public std::error_category
-{
-    virtual
-    auto
-    name() const throw() -> const char* {
-        return "cocaine.service.locator";
-    }
-
-    virtual
-    auto
-    message(int code) const -> std::string {
-        switch(code) {
-          case cocaine::error::locator_errors::service_not_available:
-            return "service is not available";
-          case cocaine::error::locator_errors::routing_storage_error:
-            return "routing storage is unavailable";
-          case cocaine::error::locator_errors::missing_version_error:
-            return "missing protocol version";
-        }
-
-        return "cocaine.service.locator error";
-    }
-};
-
-} // namespace
-
-namespace cocaine { namespace error {
-
-auto
-locator_category() -> const std::error_category& {
-    static locator_category_t instance;
-    return instance;
-}
-
-auto
-make_error_code(locator_errors code) -> std::error_code {
-    return std::error_code(static_cast<int>(code), locator_category());
-}
-
-}} // namespace cocaine::error
